+++ mstrio/api/reports.py
@@ -57,6 +57,7 @@ def report_definition(connection: 'Connection', report_id: str) -> 'Response':
 def report_instance(
     connection: 'Connection',
     report_id: str,
+    project_id: str | None = None,
     body: dict | None = None,
     offset: int = 0,
     limit: int = 5000,
@@ -91,9 +92,13 @@ def report_instance(
         params['fields'] = CUBE_FIELDS
     if execution_stage:
         params['executionStage'] = execution_stage
+    headers = {'X-MSTR-ProjectID': project_id} if project_id else {}
 
     return connection.post(
-        endpoint=f'/api/v2/reports/{report_id}/instances/', json=body, params=params
+        endpoint=f'/api/v2/reports/{report_id}/instances/',
+        json=body,
+        params=params,
+        headers=headers,
     )
 
 
@@ -376,4 +381,83 @@ def get_report_status(
         endpoint=endpoint,
         headers=headers,
         params={'fields': fields},
+    )
+
+
+@ErrorHandler(err_msg="Error getting VLDB settings for report with ID: {id}")
+def get_vldb_settings(
+    connection: 'Connection',
+    id: str,
+    instance_id: str | None = None,
+    error_msg: str | None = None,
+):
+    """Get advanced VLDB settings for a report.
+
+    Args:
+        connection (Connection): Strategy One REST API connection object
+        id (string): Datasource ID
+        instance_id (string, optional): Report Instance ID
+        error_msg (string, optional): Custom Error Message for Error Handling
+
+    Returns:
+        Complete HTTP response object.
+    """
+
+    return connection.get(
+        endpoint=f'/api/model/reports/{id}?showAdvancedProperties=true',
+        headers={'X-MSTR-MS-Instance': instance_id},
+    )
+
+
+@ErrorHandler(err_msg="Error updating VLDB settings for report with ID {id}")
+def update_vldb_settings(
+    connection: 'Connection',
+    id: str,
+    body: dict,
+    instance_id: str,
+    project_id: str | None = None,
+    error_msg: str | None = None,
+):
+    """Update metadata of advanced VLDB settings for a report.
+
+    Args:
+        connection (Connection): Strategy One REST API connection object
+        id (string): Report ID
+        instance_id (string): Report Instance ID
+        body (dict): JSON-formatted data used to update VLDB settings
+        error_msg (string, optional): Custom Error Message for Error Handling
+
+    Returns:
+        Complete HTTP response object.
+    """
+
+    return connection.put(
+        endpoint=f'/api/model/reports/{id}',
+        headers={'X-MSTR-MS-Instance': instance_id, 'X-MSTR-ProjectID': project_id},
+        json=body,
+    )
+
+
+@ErrorHandler(err_msg="Error getting metadata of VLDB settings for report with ID {id}")
+def get_applicable_vldb_settings(
+    connection: 'Connection',
+    id: str,
+    instance_id: str | None = None,
+    error_msg: str | None = None,
+):
+    """Get metadata of advanced VLDB settings for a report.
+
+    Args:
+        connection (Connection): Strategy One REST API connection object
+        id (string): Report ID
+        instance_id (string, optional): Report Instance ID
+        error_msg (string, optional): Custom Error Message for Error Handling
+
+    Returns:
+        Complete HTTP response object.
+    """
+
+    return connection.get(
+        endpoint=f'/api/model/reports/{id}/applicableVldbProperties',
+        headers={'X-MSTR-MS-Instance': instance_id},
     )